Use a sys module for OS specific information
Update cbindgen to parse modules with a path attribute and revert the
following commits:
 - 996fad709238e5c4e24395f1c4ff8c90dfb984e5.
 - cfe98ab3b2ef0d211e092c846a80b969fad82f92.
